pyproject.toml
[tool.poetry]
name = "ai-project-generator"
version = "0.1.0"
description = ""
authors = ["floris <floris@droppert.dev>"]
readme = "README.md"
packages = [{include = "ai_project_generator"}]

[tool.poetry.dependencies]
python = "^3.11"
mypy = "^1.2.0"
black = "^23.3.0"
isort = "^5.12.0"
openai = "^0.27.4"
pre-commit = "^3.2.2"
pyupgrade = "^3.3.1"
flake8 = "^6.0.0"
pytest = "^7.2.2"
pydantic = "^1.10.7"
weaviate-client = "^3.16.0"
numpy = "^1.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
